# backend/src/main/resources/openapi/tracex-api.yaml
openapi: 3.0.1
info:
  title: OpenAPI definition
  version: v0
servers:
- url: http://localhost:62436
  description: Generated server url
paths:
  /api/v1/qr/scan:
    post:
      tags:
      - qr-controller
      operationId: recordScan
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ScanRequestDto'
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ApiResponseVoid'
  /api/v1/inspections:
    get:
      tags:
      - inspection-controller
      operationId: listInspections
      parameters:
      - name: page
        in: query
        required: false
        schema:
          type: integer
          format: int32
          default: 1
      - name: limit
        in: query
        required: false
        schema:
          type: integer
          format: int32
          default: 30
      - name: status
        in: query
        required: false
        schema:
          type: string
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ApiResponseInspectionPageResult'
    post:
      tags:
      - inspection-controller
      operationId: createInspection
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/InspectionCreateDto'
      responses:
        "201":
          description: Created
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ApiResponseInspection'
...
  /api/v1/auth/users/{id}/toggle:
    patch:
      tags:
      - user-controller
      operationId: toggleUserStatus
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ApiResponseUserSummaryDto'
  /api/v1/auth/users/{id}/role:
    patch:
      tags:
      - user-controller
      operationId: updateUserRole
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateRoleDto'
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ApiResponseUserSummaryDto'
  /api/v1/auth/users/{id}/restore:
    patch:
      tags:
      - user-controller
      operationId: restoreUser
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ApiResponseUserSummaryDto'
...
  /api/v1/qr/trace/t/{token}:
    get:
      tags:
      - qr-controller
      operationId: getPublicTrace
      parameters:
      - name: token
        in: path
        required: true
        schema:
          type: string
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ApiResponsePublicTraceDto'
...
  /api/v1/inspections/{id}:
    get:
      tags:
      - inspection-controller
      operationId: getById
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ApiResponseInspection'
  /api/v1/inspections/my:
    get:
      tags:
      - inspection-controller
      operationId: listMyInspections
      parameters:
      - name: page
        in: query
        required: false
        schema:
          type: integer
          format: int32
          default: 1
      - name: limit
        in: query
        required: false
        schema:
          type: integer
          format: int32
          default: 20
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ApiResponseInspectionPageResult'
  /api/v1/inspections/batch/{batchId}:
    get:
      tags:
      - inspection-controller
      operationId: listByBatch
      parameters:
      - name: batchId
        in: path
        required: true
        schema:
          type: string
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ApiResponseListInspection'
...
  /api/v1/auth/users/{id}:
    delete:
      tags:
      - user-controller
      operationId: deleteUser
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DeleteUserDto'
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ApiResponseVoid'
...
components:
  schemas:
    ScanRequestDto:
      required:
      - token
      type: object
      properties:
        token:
          type: string
        source:
          type: string
    ApiResponseVoid:
      type: object
      properties:
        success:
          type: boolean
        requestId:
          type: string
        data:
          type: object
        message:
          type: string
        code:
          type: string
        error:
          type: string
        fieldErrors:
          type: array
          items:
            $ref: '#/components/schemas/FieldErrorDto'
    FieldErrorDto:
      type: object
      properties:
        field:
          type: string
        message:
          type: string
    ChecklistItem:
      type: object
      properties:
        label:
          type: string
        passed:
          type: boolean
        note:
          type: string
    InspectionCreateDto:
      type: object
      properties:
        batchId:
          type: string
        status:
          type: string
        rating:
          type: integer
          format: int32
        checklist:
          type: array
          items:
            $ref: '#/components/schemas/ChecklistItem'
        findings:
          type: string
        recommendation:
          type: string
        isLatest:
          type: boolean
    ApiResponseInspection:
      type: object
      properties:
        success:
          type: boolean
        requestId:
          type: string
        data:
          $ref: '#/components/schemas/Inspection'
        message:
          type: string
        code:
          type: string
        error:
          type: string
        fieldErrors:
          type: array
          items:
            $ref: '#/components/schemas/FieldErrorDto'
    InspectedBy:
      type: object
      properties:
        userId:
          type: string
        name:
          type: string
        username:
          type: string
    Inspection:
      type: object
      properties:
        id:
          type: string
        batchId:
          type: string
        batchCode:
          type: string
        productName:
          type: string
        sku:
          type: string
        status:
          type: string
        rating:
          type: integer
          format: int32
        checklist:
          type: array
          items:
            $ref: '#/components/schemas/ChecklistItem'
        findings:
          type: string
        recommendation:
          type: string
        inspectedBy:
          $ref: '#/components/schemas/InspectedBy'
        createdAt:
          type: string
          format: date-time
        isLatest:
          type: boolean
...
    UserSummaryDto:
      type: object
      properties:
        id:
          type: string
        username:
          type: string
        name:
          type: string
        email:
          type: string
        phone:
          type: string
        role:
          type: string
          enum:
          - admin
          - manager
          - factory-manager
          - quality-inspector
          - dispatch-coordinator
        emailVerified:
          type: boolean
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        active:
          type: boolean
        superAdmin:
          type: boolean
        deleted:
          type: boolean
...
    ApiResponseUserSummaryDto:
      type: object
      properties:
        success:
          type: boolean
        requestId:
          type: string
        data:
          $ref: '#/components/schemas/UserSummaryDto'
        message:
          type: string
        code:
          type: string
        error:
          type: string
        fieldErrors:
          type: array
          items:
            $ref: '#/components/schemas/FieldErrorDto'
    UpdateRoleDto:
      required:
      - role
      type: object
      properties:
        role:
          type: string
          enum:
          - admin
          - manager
          - factory-manager
          - quality-inspector
          - dispatch-coordinator
...
    ApiResponsePublicTraceDto:
      type: object
      properties:
        success:
          type: boolean
        requestId:
          type: string
        data:
          $ref: '#/components/schemas/PublicTraceDto'
        message:
          type: string
        code:
          type: string
        error:
          type: string
        fieldErrors:
          type: array
          items:
            $ref: '#/components/schemas/FieldErrorDto'
    PublicTraceDto:
      type: object
      properties:
        batchCode:
          type: string
        productName:
          type: string
        sku:
          type: string
        village:
          type: string
        packDate:
          type: string
          format: date
          example: 2026-09-15
        expiryDate:
          type: string
          format: date
          nullable: true
          example: 2026-12-15
        status:
          type: string
        qualityCheck:
          $ref: '#/components/schemas/QualityCheckPublicDto'
        traceabilityNote:
          type: string
    QualityCheckPublicDto:
      type: object
      properties:
        status:
          type: string
        rating:
          type: integer
          format: int32
        inspectedAt:
          type: string
          format: date-time
...
    ApiResponseInspectionPageResult:
      type: object
      properties:
        success:
          type: boolean
        requestId:
          type: string
        data:
          $ref: '#/components/schemas/InspectionPageResult'
        message:
          type: string
        code:
          type: string
        error:
          type: string
        fieldErrors:
          type: array
          items:
            $ref: '#/components/schemas/FieldErrorDto'
    InspectionPageResult:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: '#/components/schemas/Inspection'
        total:
          type: integer
          format: int64
        page:
          type: integer
          format: int32
        limit:
          type: integer
          format: int32
        count:
          type: integer
          format: int32
    ApiResponseListInspection:
      type: object
      properties:
        success:
          type: boolean
        requestId:
          type: string
        data:
          type: array
          items:
            $ref: '#/components/schemas/Inspection'
        message:
          type: string
        code:
          type: string
        error:
          type: string
        fieldErrors:
          type: array
          items:
            $ref: '#/components/schemas/FieldErrorDto'
...
    DeleteUserDto:
      type: object
      properties:
        deleteNote:
          type: string